requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
pyyaml
redis
taskiq[redis]  # Или celery[redis]
openai
langgraph
pydantic>=2.0
python-json-logger